more NSUInteger work. Fix nodeContentsDidChange so that if an inserIndex is ongoing, it is used as a start for all files, not just the first.

+++ GWorkspace/Desktop/GWDesktopView.h
@@ -40,8 +40,8 @@ typedef enum BackImageStyle {
 {
   NSRect screenFrame;  
   NSRect *grid;
-  int gridcount;
-  int rowcount;
+  NSInteger gridcount;
+  NSInteger rowcount;
 
   NSImage *dragIcon;
   NSPoint dragPoint;
